Refactor shared validators and converters
